+++ src/main/java/subway/handler/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.handler;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
